Separate handling blocks and texts into 2 classes
